time decode only

+++ lib/utils/unsafe_import.py
@@ -39,6 +39,6 @@ def model_from_hf_path(path,
         torch_dtype='auto',
         low_cpu_mem_usage=True,
         attn_implementation='sdpa',
-        device_map=device_map).half()
+        device_map=device_map)
 
     return model, model_str
